Use jwilder docker-gen (>=0.5.0).
Now that jwilder includes (docker >= 1.8.0 )mounts in the runtime
container, use his supported version. Note that this is a BREAKING
CHANGE, as the cyberCode version stores the mounts and volumes in the
same runtime container data structure (Volumes) and the jwilder version
stores them separately. This means that templates developed for docker <
1.8.0 cannot be reused as-is.

Also build static musl forego instead of glibc. Approach taken from:
 https://github.com/ddollar/forego/issues/65#issuecomment-146521766